Docs tweaks and handle rename of errors
